chore: move http models
