+++ http/models.go
@@ -1,4 +1,3 @@
-// TODO: move to http/models.go
 package http
 
 type ErrorResponse struct {
